+++ backend/src/main/java/com/pet/shop/services/ThuCungService.java
@@ -90,7 +90,7 @@ public class ThuCungService {
         SanPhamDTO dto = new SanPhamDTO();
         SanPham sanPham = thuCung.getSanPham();
 
-        dto.setMaSanPham(Math.toIntExact(sanPham.getMaSanPham()));
+        dto.setMaSanPham((long) Math.toIntExact(sanPham.getMaSanPham()));
         dto.setTenSanPham(sanPham.getTenSanPham());
         dto.setMaDanhMuc(Math.toIntExact(sanPham.getDanhMuc().getMaDanhMuc()));
         dto.setTenDanhMuc(sanPham.getDanhMuc().getTenDanhMuc());
